Creando vistas de los tipos de clases

+++ app/Http/Controllers/EquipoController.php
@@ -89,6 +89,7 @@ class EquipoController extends Controller
 
     public function destroy(Equipo $equipo)
     {
+        $this->authorize('delete',$equipo);
         $equipo->delete();
 
         return redirect()->route('equipos')
